Add strict validation
IPIF validation (via JSONSchema) is not strict about
additional properties.
This might be a flexibel approach, but is cumbersome from
a developers perspective for several reasons.

This is why I added strict validation (fails on additional object
properties) to the validate_factoids.py script.
Strict validation is activated by default, because in this mode
it will detect problems, which might be overseen easily.
If necessary, strict validation can be turned off via command
line option. (This implies that the user is aware of additional
properties).

The Papilotte server (at this time only with the json connector)
is still running in non-strict mode by default. This can be changed
via the configuration option `strict_validation`.

I also added a new configuration option `validate_respones`
(default to False), which will run schema validation on response
data.